Extract IdentityQueue. (#2418)
